requirements.txt
kivy
yt-dlp
arabic_reshaper
python-bidi
